modify #10 src配下にプロジェクトを移動し不要なファイルを削除
