+++ package.json
@@ -8,7 +8,7 @@
   "author": "",
   "license": "MIT",
   "dependencies": {
-    "aws-sdk": "^2.6.7",
-    "uuid": "^2.0.3"
+    "aws-sdk": "^2.7.19",
+    "uuid": "^3.0.1"
   }
 }
