--- serverless-crud/.gitignore
@@ -1,18 +1,0 @@
-.serverless
-*.log
-npm-debug.log
-pids
-*.pid
-*.seed
-dist
-lib-cov
-coverage
-.grunt
-.lock-wscript
-node_modules
-**/.idea
-.DS_Store
-.tmp
-admin.env
-.env
-tmp
