--- serverless-crud/handler.js
@@ -1,77 +1,0 @@
-'use strict';
-
-const todosCreate = require('./todos-create.js');
-const todosReadAll = require('./todos-read-all.js');
-const todosReadOne = require('./todos-read-one.js');
-const todosUpdate = require('./todos-update.js');
-const todosDelete = require('./todos-delete.js');
-
-module.exports.create = (event, context, callback) => {
-  todosCreate(event, (error, result) => {
-    const response = {
-      statusCode: 200,
-      headers: {
-        "Access-Control-Allow-Origin" : "*"
-      },
-      body: JSON.stringify(result),
-    };
-
-    context.succeed(response);
-  });
-};
-
-module.exports.readAll = (event, context, callback) => {
-  todosReadAll(event, (error, result) => {
-    const response = {
-      statusCode: 200,
-      headers: {
-        "Access-Control-Allow-Origin" : "*"
-      },
-      body: JSON.stringify(result),
-    };
-
-    context.succeed(response);
-  });
-};
-
-module.exports.readOne = (event, context, callback) => {
-  todosReadOne(event, (error, result) => {
-    const response = {
-      statusCode: 200,
-      headers: {
-        "Access-Control-Allow-Origin" : "*"
-      },
-      body: JSON.stringify(result),
-    };
-
-    context.succeed(response);
-  });
-};
-
-module.exports.update = (event, context, callback) => {
-  todosUpdate(event, (error, result) => {
-    const response = {
-      statusCode: 200,
-      headers: {
-        "Access-Control-Allow-Origin" : "*"
-      },
-      body: JSON.stringify(result),
-    };
-
-    context.succeed(response);
-  });
-};
-
-module.exports.delete = (event, context, callback) => {
-  todosDelete(event, (error, result) => {
-    const response = {
-      statusCode: 200,
-      headers: {
-        "Access-Control-Allow-Origin" : "*"
-      },
-      body: JSON.stringify(result),
-    };
-
-    context.succeed(response);
-  });
-};
