--- serverless-crud/serverless.yml
@@ -1,74 +1,0 @@
-service: serverless-crud
-
-provider:
-  name: aws
-  runtime: nodejs4.3
-  region: ap-northeast-1
-  stage: dev
-  iamRoleStatements:
-    - Effect: Allow
-      Action:
-        - dynamodb:DescribeTable
-        - dynamodb:Query
-        - dynamodb:Scan
-        - dynamodb:GetItem
-        - dynamodb:PutItem
-        - dynamodb:UpdateItem
-        - dynamodb:DeleteItem
-      Resource: "arn:aws:dynamodb:ap-northeast-1:*:*"
-
-functions:
-  create:
-    handler: handler.create
-    events:
-      - http:
-          path: todos
-          method: post
-          cors: true
-  readAll:
-    handler: handler.readAll
-    events:
-      - http:
-          path: todos
-          method: get
-          cors: true
-  readOne:
-    handler: handler.readOne
-    events:
-      - http:
-          path: todos/{id}
-          method: get
-          cors: true
-  update:
-    handler: handler.update
-    events:
-      - http:
-          path: todos/{id}
-          method: put
-          cors: true
-  delete:
-    handler: handler.delete
-    events:
-      - http:
-          path: todos/{id}
-          method: delete
-          cors: true
-
-resources:
-  Resources:
-    TodosDynamoDbTable:
-      Type: 'AWS::DynamoDB::Table'
-      DeletionPolicy: Retain
-      Properties:
-        AttributeDefinitions:
-          -
-            AttributeName: id
-            AttributeType: S
-        KeySchema:
-          -
-            AttributeName: id
-            KeyType: HASH
-        ProvisionedThroughput:
-          ReadCapacityUnits: 1
-          WriteCapacityUnits: 1
-        TableName: 'todos'
