--- serverless-crud/todos-delete.js
@@ -1,20 +1,0 @@
-'use strict';
-
-const AWS = require('aws-sdk');
-const dynamoDb = new AWS.DynamoDB.DocumentClient();
-
-module.exports = (event, callback) => {
-  const params = {
-    TableName : 'todos',
-    Key: {
-      id: event.pathParameters.id
-    }
-  };
-
-  return dynamoDb.delete(params, (error, data) => {
-    if (error) {
-      callback(error);
-    }
-    callback(error, params.Key);
-  });
-};
