--- serverless-crud/todos-read-all.js
@@ -1,17 +1,0 @@
-'use strict';
-
-const AWS = require('aws-sdk');
-const dynamoDb = new AWS.DynamoDB.DocumentClient();
-
-module.exports = (event, callback) => {
-  const params = {
-    TableName: 'todos',
-  };
-
-  return dynamoDb.scan(params, (error, data) => {
-    if (error) {
-      callback(error);
-    }
-    callback(error, data.Items);
-  });
-};
