+++ src/handler.js
@@ -1,0 +1,32 @@
+'use strict';
+
+const todosCreate = require('./todos-create.js');
+const todosReadOne = require('./todos-read-one.js');
+
+module.exports.create = (event, context, callback) => {
+  todosCreate(event, (error, result) => {
+    const response = {
+      statusCode: 200,
+      headers: {
+        "Access-Control-Allow-Origin" : "*"
+      },
+      body: JSON.stringify(result),
+    };
+
+    context.succeed(response);
+  });
+};
+
+module.exports.readOne = (event, context, callback) => {
+  todosReadOne(event, (error, result) => {
+    const response = {
+      statusCode: 200,
+      headers: {
+        "Access-Control-Allow-Origin" : "*"
+      },
+      body: JSON.stringify(result),
+    };
+
+    context.succeed(response);
+  });
+};
